bzr revid: tde@openerp.com-20140418155417-lho910ewzfp3556o

+++ addons/website_forum/controllers/main.py
@@ -32,18 +32,35 @@ class WebsiteForum(http.Controller):
         return msg
 
     def _prepare_forum_values(self, forum=None, **kwargs):
+        Forum = request.registry['forum.forum']
         user = request.registry['res.users'].browse(request.cr, request.uid, request.uid, context=request.context)
         public_uid = request.registry['website'].get_public_user(request.cr, request.uid, request.context)
         values = {'user': user, 'is_public_user': user.id == public_uid,
                   'notifications': self._get_notifications(),
                   'header': kwargs.get('header', dict()),
-                  'searches': kwargs.get('searches', dict())}
+                  'searches': kwargs.get('searches', dict()),
+                  'can_edit_own': True,
+                  'can_edit_all': user.karma > Forum._karma_modo_edit_all,
+                  'can_close_own': user.karma > Forum._karma_modo_close_own,
+                  'can_close_all': user.karma > Forum._karma_modo_close_all,
+                  'can_unlink_own': user.karma > Forum._karma_modo_unlink_own,
+                  'can_unlink_all': user.karma > Forum._karma_modo_unlink_all,
+                  'can_unlink_comment': user.karma > Forum._karma_modo_unlink_comment,
+                  }
         if forum:
             values['forum'] = forum
         elif kwargs.get('forum_id'):
             values['forum'] = request.registry['forum.forum'].browse(request.cr, request.uid, kwargs.pop('forum_id'), context=request.context)
         values.update(kwargs)
         return values
+
+    def _has_enough_karma(self, karma_name, uid=None):
+        Forum = request.registry['forum.forum']
+        karma = hasattr(Forum, karma_name) and getattr(Forum, karma_name) or 0
+        user = request.registry['res.users'].browse(request.cr, SUPERUSER_ID, uid or request.uid, context=request.context)
+        if user.karma < karma:
+            return False, {'error': 'not_enough_karma', 'karma': karma}
+        return True, {}
 
     # Forum
     # --------------------------------------------------
@@ -201,6 +218,10 @@ class WebsiteForum(http.Controller):
 
     @http.route('/forum/<model("forum.forum"):forum>/question/<model("forum.post"):question>/ask_for_close', type='http', auth="user", multilang=True, website=True)
     def question_ask_for_close(self, forum, question, **post):
+        check_res = self._has_enough_karma(question.create_uid.id == request.uid and '_karma_modo_close_own' or '_karma_modo_close_all')
+        if not check_res[0]:
+            return werkzeug.utils.redirect("/forum/%s" % slug(forum))
+
         cr, uid, context = request.cr, request.uid, request.context
         Reason = request.registry['forum.post.reason']
         reason_ids = Reason.search(cr, uid, [], context=context)
@@ -208,7 +229,7 @@ class WebsiteForum(http.Controller):
 
         values = self._prepare_forum_values(**post)
         values.update({
-            'post': question,
+            'question': question,
             'forum': forum,
             'reasons': reasons,
         })
@@ -224,6 +245,10 @@ class WebsiteForum(http.Controller):
 
     @http.route('/forum/<model("forum.forum"):forum>/question/<model("forum.post"):question>/close', type='http', auth="user", multilang=True, methods=['POST'], website=True)
     def question_close(self, forum, question, **post):
+        check_res = self._has_enough_karma(question.create_uid.id == request.uid and '_karma_modo_close_own' or '_karma_modo_close_all')
+        if not check_res[0]:
+            return werkzeug.utils.redirect("/forum/%s" % slug(forum))
+
         request.registry['forum.post'].write(request.cr, request.uid, [question.id], {
             'state': 'close',
             'closed_uid': request.uid,
@@ -234,17 +259,28 @@ class WebsiteForum(http.Controller):
 
     @http.route('/forum/<model("forum.forum"):forum>/question/<model("forum.post"):question>/reopen', type='http', auth="user", multilang=True, website=True)
     def question_reopen(self, forum, question, **kwarg):
+        check_res = self._has_enough_karma(question.create_uid.id == request.uid and '_karma_modo_close_own' or '_karma_modo_close_all')
+        if not check_res[0]:
+            return werkzeug.utils.redirect("/forum/%s" % slug(forum))
+
         request.registry['forum.post'].write(request.cr, request.uid, [question.id], {'state': 'active'}, context=request.context)
         return werkzeug.utils.redirect("/forum/%s/question/%s" % (slug(forum), slug(question)))
 
     @http.route('/forum/<model("forum.forum"):forum>/question/<model("forum.post"):question>/delete', type='http', auth="user", multilang=True, website=True)
     def question_delete(self, forum, question, **kwarg):
-        #instead of unlink record just change 'active' to false so user can undelete it.
+        check_res = self._has_enough_karma(question.create_uid.id == request.uid and '_karma_modo_unlink_own' or '_karma_modo_unlink_all')
+        if not check_res[0]:
+            return werkzeug.utils.redirect("/forum/%s" % slug(forum))
+
         request.registry['forum.post'].write(request.cr, request.uid, [question.id], {'active': False}, context=request.context)
         return werkzeug.utils.redirect("/forum/%s/question/%s" % (slug(forum), slug(question)))
 
     @http.route('/forum/<model("forum.forum"):forum>/question/<model("forum.post"):question>/undelete', type='http', auth="user", multilang=True, website=True)
     def question_undelete(self, forum, question, **kwarg):
+        check_res = self._has_enough_karma(question.create_uid.id == request.uid and '_karma_modo_unlink_own' or '_karma_modo_unlink_all')
+        if not check_res[0]:
+            return werkzeug.utils.redirect("/forum/%s" % slug(forum))
+
         request.registry['forum.post'].write(request.cr, request.uid, [question.id], {'active': True}, context=request.context)
         return werkzeug.utils.redirect("/forum/%s/question/%s" % (slug(forum), slug(question)))
 
@@ -283,19 +319,27 @@ class WebsiteForum(http.Controller):
     @http.route('/forum/<model("forum.forum"):forum>/post/<model("forum.post"):post>/toggle_correct', type='json', auth="public", website=True)
     def post_toggle_correct(self, forum, post, **kwargs):
         cr, uid, context = request.cr, request.uid, request.context
+        if post.parent_id is False:
+            return request.redirect('/')
         if not request.session.uid:
             return {'error': 'anonymous_user'}
-        # if user have not access to accept answer then reise warning
-        if post.parent_id is False or post.parent_id.create_uid.id != uid:
+        user = request.registry['res.users'].browse(request.cr, SUPERUSER_ID, request.uid, context=request.context)
+        if post.parent_id.create_uid.id != uid:
             return {'error': 'own_post'}
+        if post.create_uid.id == user.id and user.karma < request.registry['forum.forum']._karma_answer_accept_own:
+            return {'error': 'not_enough_karma', 'karma': 20}
 
         # set all answers to False, only one can be accepted
         request.registry['forum.post'].write(cr, uid, [c.id for c in post.parent_id.child_ids], {'is_correct': False}, context=context)
-        request.registry['forum.post'].write(cr, uid, [post.id, post.parent_id.id], {'is_correct': not post.is_correct}, context=context)
+        request.registry['forum.post'].write(cr, uid, [post.id], {'is_correct': not post.is_correct}, context=context)
         return not post.is_correct
 
     @http.route('/forum/<model("forum.forum"):forum>/post/<model("forum.post"):post>/delete', type='http', auth="user", multilang=True, website=True)
     def post_delete(self, forum, post, **kwargs):
+        check_res = self._has_enough_karma(post.create_uid.id == request.uid and '_karma_modo_unlink_own' or '_karma_modo_unlink_all')
+        if not check_res[0]:
+            return werkzeug.utils.redirect("/forum/%s" % slug(forum))
+
         question = post.parent_id
         request.registry['forum.post'].unlink(request.cr, request.uid, [post.id], context=request.context)
         if question:
@@ -304,6 +348,10 @@ class WebsiteForum(http.Controller):
 
     @http.route('/forum/<model("forum.forum"):forum>/post/<model("forum.post"):post>/edit', type='http', auth="user", website=True, multilang=True)
     def post_edit(self, forum, post, **kwargs):
+        check_res = self._has_enough_karma(post.create_uid.id == request.uid and '_karma_modo_edit_own' or '_karma_modo_edit_all')
+        if not check_res[0]:
+            return werkzeug.utils.redirect("/forum/%s" % slug(forum))
+
         tags = ""
         for tag_name in post.tag_ids:
             tags += tag_name.name + ","
@@ -341,15 +389,15 @@ class WebsiteForum(http.Controller):
 
     @http.route('/forum/<model("forum.forum"):forum>/post/<model("forum.post"):post>/upvote', type='json', auth="public", multilang=True, website=True)
     def post_upvote(self, forum, post, **kwargs):
-        # check for karma and not self vote
         if not request.session.uid:
             return {'error': 'anonymous_user'}
         if request.uid == post.create_uid.id:
             return {'error': 'own_post'}
-        user = request.registry['res.users'].browse(request.cr, SUPERUSER_ID, request.uid, context=request.context)
-        if user.karma <= 5:
-            return {'error': 'not_enough_karma', 'karma': 1}
-        return request.registry['forum.post'].vote(request.cr, request.uid, [post.id], upvote=True, context=request.context)
+        check_res = self._has_enough_karma('_karma_upvote')
+        if not check_res[0]:
+            return check_res[1]
+        upvote = True if not post.user_vote > 0 else False
+        return request.registry['forum.post'].vote(request.cr, request.uid, [post.id], upvote=upvote, context=request.context)
 
     @http.route('/forum/<model("forum.forum"):forum>/post/<model("forum.post"):post>/downvote', type='json', auth="public", multilang=True, website=True)
     def post_downvote(self, forum, post, **kwargs):
@@ -357,10 +405,11 @@ class WebsiteForum(http.Controller):
             return {'error': 'anonymous_user'}
         if request.uid == post.create_uid.id:
             return {'error': 'own_post'}
-        user = request.registry['res.users'].browse(request.cr, SUPERUSER_ID, request.uid, context=request.context)
-        if user.karma <= 50:
-            return {'error': 'not_enough_karma', 'karma': 50}
-        return request.registry['forum.post'].vote(request.cr, request.uid, [post.id], upvote=False, context=request.context)
+        check_res = self._has_enough_karma('_karma_downvote')
+        if not check_res[0]:
+            return check_res[1]
+        upvote = True if post.user_vote < 0 else False
+        return request.registry['forum.post'].vote(request.cr, request.uid, [post.id], upvote=upvote, context=request.context)
 
     # User
     # --------------------------------------------------
@@ -386,6 +435,16 @@ class WebsiteForum(http.Controller):
         })
 
         return request.website.render("website_forum.users", values)
+
+    @http.route(['/forum/<model("forum.forum"):forum>/partner/<int:partner_id>'], type='http', auth="public", website=True, multilang=True)
+    def open_partner(self, forum, partner_id=0, **post):
+        cr, uid, context = request.cr, request.uid, request.context
+        pids = request.registry['res.partner'].search(cr, SUPERUSER_ID, [('id', '=', partner_id)], context=context)
+        if pids:
+            partner = request.registry['res.partner'].browse(cr, SUPERUSER_ID, pids[0], context=context)
+            if partner.user_ids:
+                return werkzeug.utils.redirect("/forum/%s/user/%d" % (slug(forum), partner.user_ids[0].id))
+        return werkzeug.utils.redirect("/forum/%s" % slug(forum))
 
     @http.route(['/forum/<model("forum.forum"):forum>/user/<int:user_id>'], type='http', auth="public", website=True, multilang=True)
     def open_user(self, forum, user_id=0, **post):
